requirements.txt
# Data Collection
praw>=7.7.1  # Reddit API
pandas>=2.0.0
numpy>=1.24.0

# Sentiment Analysis
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0 

# Data Processing
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

